+++ packages/expo-app-metrics/CHANGELOG.md
@@ -13,3 +13,5 @@
 - fix race condition between db inserts ([#46702](https://github.com/expo/expo/pull/46702) by [@Ubax](https://github.com/Ubax))
 
 ### 💡 Others
+
+- Add private `getForegroundSession` ([#46657](https://github.com/expo/expo/pull/46657) by [@Ubax](https://github.com/Ubax))